refactor(langy/sidebar): extract hooks to useLangyShortcuts.ts (rule 11)
LangySidebar.tsx is 1453 lines mixing the LangyDrawer entry point,
the LangyPanel and ~10 of its tightly-coupled sub-components
(PanelHeader, RecentList, EmptyState, Chip, ThinkingIndicator,
Composer, SendButton, MessageContent, ProposalCard), four sparkle
visuals, the LangyHandle, two hooks, and several shared constants
and types.

This commit does the SAFE extraction only — the two hooks
(useTypewriterPlaceholder, useGlobalLangyShortcut) move to
`useLangyShortcuts.ts` since they're self-contained (no React state
shared with the rest of the file, just their own state + an effect)
and the move can be verified by grep alone.

LangySidebar.tsx is now 1356 lines (-97). Still over the rule-11 bar
but materially smaller.

Deferred to follow-up PRs: extracting ProposalCard (~205 lines),
LangyHandle (~80), the sparkle/gradient visuals (~105), and ideally
splitting LangyPanel's inner sub-components into LangyComposer +
LangyMessageContent + LangyRecentList. Each of those requires
threading shared constants (PANEL_WIDTH, AI_SHADOW, AI_BG_SUBTLE,
SPARKLE_GRADIENT_ID, etc.) through their new locations and is
worth a separate review-able PR with local typecheck verification —
something this worktree doesn't have install state for.
